+++ R/print.check_model.R
@@ -1,4 +1,4 @@
-#' @param checks Character vector, indicating which checks for \code{performance::check_model()}
+#' @param check Character vector, indicating which checks for \code{performance::check_model()}
 #'   should be performed and plotted. May be one or more of
 #'   \code{"all", "vif", "qq", "normality", "ncv", "homogeneity", "outliers", "reqq"}.
 #'   \code{"reqq"} is a QQ-plot for random effects and only available for mixed models.
@@ -8,7 +8,7 @@
 #' @importFrom graphics plot
 #' @rdname data_plot
 #' @export
-print.see_check_model <- function(x, checks = "all", ...) {
+print.see_check_model <- function(x, check = "all", ...) {
   p <- list()
 
   panel <- attr(x, "panel")
@@ -16,14 +16,14 @@ print.see_check_model <- function(x, checks = "all", ...) {
   line_size <- attr(x, "line_size")
   text_size <- attr(x, "text_size")
 
-  checks <- tolower(checks)
+  check <- tolower(check)
 
-  if ("VIF" %in% names(x) && c("vif", "all") %in% checks) p$VIF <- .plot_diag_vif(x$VIF)
-  if ("QQ" %in% names(x) && c("qq", "all") %in% checks) p$QQ <- .plot_diag_qq(x$QQ, dot_size, line_size)
-  if ("NORM" %in% names(x) && c("normality", "all") %in% checks) p$NORM <- .plot_diag_norm(x$NORM, line_size)
-  if ("NCV" %in% names(x) && c("ncv", "all") %in% checks) p$NCV <- .plot_diag_ncv(x$NCV, dot_size, line_size)
-  if ("HOMOGENEITY" %in% names(x) && c("homogeneity", "all") %in% checks) p$HOMOGENEITY <- .plot_diag_homogeneity(x$HOMOGENEITY, dot_size, line_size)
-  if ("OUTLIERS" %in% names(x) && c("outliers", "all") %in% checks) {
+  if ("VIF" %in% names(x) && c("vif", "all") %in% check) p$VIF <- .plot_diag_vif(x$VIF)
+  if ("QQ" %in% names(x) && c("qq", "all") %in% check) p$QQ <- .plot_diag_qq(x$QQ, dot_size, line_size)
+  if ("NORM" %in% names(x) && c("normality", "all") %in% check) p$NORM <- .plot_diag_norm(x$NORM, line_size)
+  if ("NCV" %in% names(x) && c("ncv", "all") %in% check) p$NCV <- .plot_diag_ncv(x$NCV, dot_size, line_size)
+  if ("HOMOGENEITY" %in% names(x) && c("homogeneity", "all") %in% check) p$HOMOGENEITY <- .plot_diag_homogeneity(x$HOMOGENEITY, dot_size, line_size)
+  if ("OUTLIERS" %in% names(x) && c("outliers", "all") %in% check) {
     p$OUTLIERS <- .plot_diag_outliers(x$OUTLIERS, text_size)
     p$OUTLIERS <- p$OUTLIERS +
       theme_lucid(
@@ -33,7 +33,7 @@ print.see_check_model <- function(x, checks = "all", ...) {
       )
   }
 
-  if ("REQQ" %in% names(x) && c("reqq", "all") %in% checks) {
+  if ("REQQ" %in% names(x) && c("reqq", "all") %in% check) {
     ps <- .plot_diag_reqq(x$REQQ, dot_size, line_size)
     for (i in 1:length(ps)) {
       p[[length(p) + 1]] <- ps[[i]]